Revert "Changed Sum value to double from int. Should probably be able to make floating-point transactions."
This reverts commit f573aa2092469af43ce544f8ce319be0cf2a01f3.

+++ src/main/java/se/chalmers/bokforing/model/Sum.java
@@ -24,12 +24,12 @@ public class Sum implements Serializable {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private int id;
     
-    private double value;
+    private int value;
     
     @Enumerated(EnumType.STRING)
     private PostType type;
     
-    public Sum(double value, PostType type){
+    public Sum(int value, PostType type){
         this.value = value;
         this.type = type;
     }
